+++ spec/controllers/posts_controller_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 describe PostsController, type: :controller do
   include_context 'JSON API headers'
