+++ spec/features/record_count_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 ##
 # Configs
