+++ spec/rails_helper.rb
@@ -1,0 +1,18 @@
+require 'spec_helper'
+
+require 'rails/all'
+require 'rails/test_help'
+require 'rspec/rails'
+
+require 'jsonapi-resources'
+require 'jsonapi/utils'
+
+require 'support/models'
+require 'support/factories'
+require 'support/resources'
+require 'support/controllers'
+require 'support/paginators'
+require 'support/application'
+
+require 'support/shared/jsonapi_errors'
+require 'support/shared/jsonapi_request'
